Only Install Steam If Needed
To allow this script to be run twice, only use apt update and apt
install for steam and other components at the first run.

Please note that this will not cover if a users installed steam manually
but did not install steam-devices or x11-utils. Skipping this case here
since we assume that this script is run on vanilla ubuntu.

+++ install.sh
@@ -129,9 +129,11 @@ if [[ "${INCLUDE_GPU_DRIVERS}" == "true" ]]; then
 fi
 
 # Install steam and steam device support.
-echo "Installing steam..."
-apt update
-apt install steam steam-devices x11-utils -y
+if [ ! -e /usr/games/steam ]; then
+	echo "Installing steam..."
+	apt update
+	apt install steam steam-devices x11-utils -y
+fi
 
 # WIP - find a way to enable Steamplay without using Desktop Steam Client. Also maybe find a way to enable Steam Beta with latest Steamplay
 # Enable SteamPlay
